グローバルエラハンドラーをReact Queryから切り離して初期化処理でも使用できるよう修正
