+++ example-app/SantokuApp/src/framework/error/useDefaultGlobalErrorHandler.test.tsx
@@ -4,15 +4,13 @@ import {useSnackbar, WithSnackbar} from 'components/overlay';
 import {loadBundledMessagesAsync} from 'framework/initialize/helpers';
 import React from 'react';
 import {Alert} from 'react-native';
-import {Mutation, Query, QueryKey} from 'react-query';
 
-import {useDefaultGlobalMutationErrorHandler, useDefaultGlobalQueryErrorHandler} from './useDefaultGlobalErrorHandler';
+import {useDefaultGlobalErrorHandler} from './useDefaultGlobalErrorHandler';
 
 jest.mock('components/overlay/snackbar/WithSnackbar');
 jest.mock('framework/logging');
 
-describe('useDefaultGlobalQueryErrorHandler', () => {
-  const query = {} as unknown as Query<unknown, unknown, unknown, QueryKey>;
+describe('useDefaultGlobalErrorHandler', () => {
   const mockSnackbarShow = jest.fn();
 
   beforeAll(() => {
@@ -36,9 +34,9 @@ describe('useDefaultGlobalQueryErrorHandler', () => {
       toJSON: () => {},
     } as unknown as AxiosError;
     await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
+    const {result: errorHandler} = renderHook(() => useDefaultGlobalErrorHandler());
     expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
+    errorHandler.current(axiosError);
     expect(mockSnackbarShow).not.toBeCalled();
   });
 
@@ -51,9 +49,9 @@ describe('useDefaultGlobalQueryErrorHandler', () => {
     } as unknown as AxiosError;
     const spyAlert = jest.spyOn(Alert, 'alert');
     await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
+    const {result: errorHandler} = renderHook(() => useDefaultGlobalErrorHandler());
     expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
+    errorHandler.current(axiosError);
     expect(mockSnackbarShow).not.toBeCalled();
     expect(spyAlert).toBeCalledWith(
       '再ログインが必要です',
@@ -70,9 +68,9 @@ describe('useDefaultGlobalQueryErrorHandler', () => {
     } as unknown as AxiosError;
     const spyAlert = jest.spyOn(Alert, 'alert');
     await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
+    const {result: errorHandler} = renderHook(() => useDefaultGlobalErrorHandler());
     expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
+    errorHandler.current(axiosError);
     expect(mockSnackbarShow).not.toBeCalled();
     expect(spyAlert).toBeCalledWith(
       '新しい利用規約への同意',
@@ -88,9 +86,9 @@ describe('useDefaultGlobalQueryErrorHandler', () => {
       toJSON: () => {},
     } as unknown as AxiosError;
     await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
+    const {result: errorHandler} = renderHook(() => useDefaultGlobalErrorHandler());
     expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
+    errorHandler.current(axiosError);
     expect(mockSnackbarShow).not.toBeCalled();
   });
 
@@ -103,9 +101,9 @@ describe('useDefaultGlobalQueryErrorHandler', () => {
     } as unknown as AxiosError;
     const spyAlert = jest.spyOn(Alert, 'alert');
     await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
+    const {result: errorHandler} = renderHook(() => useDefaultGlobalErrorHandler());
     expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
+    errorHandler.current(axiosError);
     expect(mockSnackbarShow).not.toBeCalled();
     expect(spyAlert).toBeCalledWith(
       'アプリケーションの更新が必要です',
@@ -121,9 +119,9 @@ describe('useDefaultGlobalQueryErrorHandler', () => {
       toJSON: () => {},
     } as unknown as AxiosError;
     await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
+    const {result: errorHandler} = renderHook(() => useDefaultGlobalErrorHandler());
     expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
+    errorHandler.current(axiosError);
     expect(mockSnackbarShow).toBeCalledWith(
       'ただいまリクエストが集中しており混雑しております。時間をおいてから再度お試しください。',
     );
@@ -137,9 +135,9 @@ describe('useDefaultGlobalQueryErrorHandler', () => {
       toJSON: () => {},
     } as unknown as AxiosError;
     await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
+    const {result: errorHandler} = renderHook(() => useDefaultGlobalErrorHandler());
     expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
+    errorHandler.current(axiosError);
     expect(mockSnackbarShow).toBeCalledWith(
       'ただいまシステムメンテナンスを実施しております。時間をおいてから再度お試しください。',
     );
@@ -153,9 +151,9 @@ describe('useDefaultGlobalQueryErrorHandler', () => {
       toJSON: () => {},
     } as unknown as AxiosError;
     await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
+    const {result: errorHandler} = renderHook(() => useDefaultGlobalErrorHandler());
     expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
+    errorHandler.current(axiosError);
     expect(mockSnackbarShow).toBeCalledWith(
       'サーバへの接続がタイムアウトしました。時間をおいてから再度お試しください。',
     );
@@ -169,63 +167,11 @@ describe('useDefaultGlobalQueryErrorHandler', () => {
       toJSON: () => {},
     } as unknown as AxiosError;
     await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
+    const {result: errorHandler} = renderHook(() => useDefaultGlobalErrorHandler());
     expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
+    errorHandler.current(axiosError);
     expect(mockSnackbarShow).toBeCalledWith(
       '予期せぬ通信エラーが発生しました。時間をおいてから再度お試しいただき、解決しない場合はお問い合わせください。',
     );
-  });
-
-  test('disableGlobalErrorHandlerが設定されている場合に何も行わない', async () => {
-    const query = {meta: {disableGlobalErrorHandler: true}} as unknown as Query<unknown, unknown, unknown, QueryKey>;
-    const axiosError = {
-      config: jest.fn(),
-      response: {status: 500, statusText: 'Internal Server Error', data: {message: 'message', code: 'errorCode'}},
-      isAxiosError: true,
-      toJSON: () => {},
-    } as unknown as AxiosError;
-    await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalQueryErrorHandler());
-    expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, query);
-    expect(mockSnackbarShow).not.toBeCalled();
-  });
-});
-
-describe('useDefaultGlobalMutationErrorHandler', () => {
-  const mockSnackbarShow = jest.fn();
-
-  beforeAll(() => {
-    (useSnackbar as jest.Mock).mockImplementation(() => ({
-      show: mockSnackbarShow,
-    }));
-    (WithSnackbar as jest.Mock).mockReturnValue(<></>);
-  });
-
-  beforeEach(() => {
-    mockSnackbarShow.mockClear();
-    (useSnackbar as jest.Mock).mockClear();
-    (WithSnackbar as jest.Mock).mockClear();
-  });
-
-  test('disableGlobalErrorHandlerが設定されている場合に何も行わない', async () => {
-    const mutation = {meta: {disableGlobalErrorHandler: true}} as unknown as Mutation<
-      unknown,
-      unknown,
-      unknown,
-      unknown
-    >;
-    const axiosError = {
-      config: jest.fn(),
-      response: {status: 500, statusText: 'Internal Server Error', data: {message: 'message', code: 'errorCode'}},
-      isAxiosError: true,
-      toJSON: () => {},
-    } as unknown as AxiosError;
-    await loadBundledMessagesAsync();
-    const {result: errorHandler} = renderHook(() => useDefaultGlobalMutationErrorHandler());
-    expect(errorHandler.current).not.toBeUndefined();
-    errorHandler.current(axiosError, jest.fn(), jest.fn(), mutation);
-    expect(mockSnackbarShow).not.toBeCalled();
   });
 });
